chore: Rework on how plugin being registered
