+++ src/main/java/io/github/null2264/cobblegen/integration/viewer/FluidInteractionRecipeHolder.java
@@ -1,4 +1,4 @@
-package io.github.null2264.cobblegen.integration;
+package io.github.null2264.cobblegen.integration.viewer;
 
 import io.github.null2264.cobblegen.config.WeightedBlock;
 import io.github.null2264.cobblegen.util.GeneratorType;
