+++ src/main/java/io/github/null2264/cobblegen/integration/viewer/emi/FluidInteractionRecipe.java
@@ -1,4 +1,4 @@
-package io.github.null2264.cobblegen.integration.emi;
+package io.github.null2264.cobblegen.integration.viewer.emi;
 
 import dev.emi.emi.api.recipe.EmiRecipe;
 import dev.emi.emi.api.recipe.EmiRecipeCategory;
@@ -6,8 +6,8 @@ import dev.emi.emi.api.stack.EmiIngredient;
 import dev.emi.emi.api.stack.EmiStack;
 import dev.emi.emi.api.widget.WidgetHolder;
 import io.github.null2264.cobblegen.config.WeightedBlock;
-import io.github.null2264.cobblegen.integration.FluidInteractionRecipeHolder;
-import io.github.null2264.cobblegen.data.Constants;
+import io.github.null2264.cobblegen.integration.viewer.FluidInteractionRecipeHolder;
+import io.github.null2264.cobblegen.util.Constants;
 import io.github.null2264.cobblegen.util.GeneratorType;
 import io.github.null2264.cobblegen.util.Util;
 import lombok.val;
