+++ src/main/java/io/github/null2264/cobblegen/integration/viewer/rei/CGREIPlugin.java
@@ -1,4 +1,4 @@
-package io.github.null2264.cobblegen.integration.rei;
+package io.github.null2264.cobblegen.integration.viewer.rei;
 
 import io.github.null2264.cobblegen.config.WeightedBlock;
 import io.github.null2264.cobblegen.util.GeneratorType;
