+++ src/main/java/io/github/null2264/cobblegen/util/CGLog.java
@@ -7,15 +7,27 @@ public class CGLog {
     public static final Logger LOG = LoggerFactory.getLogger("CobbleGen");
     private static final String prefix = "[CobbleGen] ";
 
-    public static void info(String s) {
-        LOG.info(prefix + s);
+    public static void info(String s, String... a) {
+        StringBuilder rt = new StringBuilder(prefix + s);
+        for (String s1 : a) {
+            rt.append(" ").append(s1);
+        }
+        LOG.info(rt.toString());
     }
 
-    public static void warn(String s) {
-        LOG.warn(prefix + s);
+    public static void warn(String s, String... a) {
+        StringBuilder rt = new StringBuilder(prefix + s);
+        for (String s1 : a) {
+            rt.append(" ").append(s1);
+        }
+        LOG.warn(rt.toString());
     }
 
-    public static void error(String s) {
-        LOG.error(prefix + s);
+    public static void error(String s, String... a) {
+        StringBuilder rt = new StringBuilder(prefix + s);
+        for (String s1 : a) {
+            rt.append(" ").append(s1);
+        }
+        LOG.error(rt.toString());
     }
 }
