+++ src/main/java/io/github/null2264/cobblegen/util/Constants.java
@@ -1,6 +1,5 @@
-package io.github.null2264.cobblegen.data;
+package io.github.null2264.cobblegen.util;
 
-import io.github.null2264.cobblegen.util.Util;
 import net.minecraft.util.Identifier;
 
 public class Constants
