+++ src/main/java/io/github/null2264/cobblegen/util/PluginFinder.java
@@ -1,12 +1,24 @@
 package io.github.null2264.cobblegen.util;
 
 import net.fabricmc.loader.api.FabricLoader;
+import net.fabricmc.loader.api.entrypoint.EntrypointContainer;
 
+import java.util.Comparator;
 import java.util.List;
+
+import static io.github.null2264.cobblegen.CobbleGen.MOD_ID;
 
 public class PluginFinder
 {
-    public static List<CobbleGenPlugin> getModPlugins() {
-        return FabricLoader.getInstance().getEntrypoints("cobblegen_plugin", CobbleGenPlugin.class);
+    public static List<EntrypointContainer<CobbleGenPlugin>> getModPlugins() {
+        return FabricLoader.getInstance()
+                .getEntrypointContainers("cobblegen_plugin", CobbleGenPlugin.class)
+                .stream()
+                .sorted(Comparator.comparingInt(PluginFinder::priorityEntrypoint))
+                .toList();
+    }
+
+    private static int priorityEntrypoint(EntrypointContainer<CobbleGenPlugin> plugin) {
+        return plugin.getProvider().getMetadata().getId().equals(MOD_ID) ? 0 : 1;
     }
 }
