+++ src/main/resources/fabric.mod.json
@@ -20,13 +20,16 @@
             "io.github.null2264.cobblegen.CobbleGen"
         ],
         "jei_mod_plugin": [
-            "io.github.null2264.cobblegen.integration.jei.CGJEIPlugin"
+            "io.github.null2264.cobblegen.integration.viewer.jei.CGJEIPlugin"
         ],
         "rei_client": [
-            "io.github.null2264.cobblegen.integration.rei.CGREIPlugin"
+            "io.github.null2264.cobblegen.integration.viewer.rei.CGREIPlugin"
         ],
         "emi": [
-            "io.github.null2264.cobblegen.integration.emi.CGEMIPlugin"
+            "io.github.null2264.cobblegen.integration.viewer.emi.CGEMIPlugin"
+        ],
+        "cobblegen_plugin": [
+            "io.github.null2264.cobblegen.integration.BuiltInPlugin"
         ]
     },
     "mixins": [
